Make it possible to mount raw partition.
This is useful for e.g. mounting a littlefs blob embedded in
the firmware.

Add a 'partition' field to the configuration struct,
which will be used if the partition_label is NULL.

Add a read_only field to the configuration struct
to mount the filesystem as read-only.

The filesystem is mounted read-only by marking all vfs
functions that modify the partition as NULL.

Add unittests for reading from a static filesystem.

Add tests for being unable to delete, write and
create files.